nouveau: Force iMac9,1 backlight PWM DIV to 1.
The iMac9,1 GeForce 9400 device has NV50_PDISP_SOR_PWM_DIV
set to 0x1 by default on boot. Upon resuming via the vbios
post routines the DIV will change to 0x84.
With this changed range the device does not have the same
brightness at the same percentages. It is dark as low as 50%.

Force the value to 1 to keep the behavior consistent.

Change the exposed backlight values for the iMac9,1 to
included a range from the minimum visible to the maximum
visible values. Allow the user space tools to be as
granular as they want.

OVER-3823

Link: https://bugs.freedesktop.org/show_bug.cgi?id=98677