efi: Lock kernel down with config option.
In commit 7756b3b63021 ("efi: Lock down the kernel if booted in secure boot
mode") the LOCK_DOWN_IN_EFI_SECURE_BOOT config option was added
but not used. It depends on CONFIG_SECURITY_LOCKDOWN_LSM which
is required for the lock_kernel_down symbol.
If CONFIG_SECURITY_LOCKDOWN_LSM is not set but CONFIG_EFI is set
this will fail to compile.